+++ octgnFX/Octgn/Networking/ClientHandler.cs
@@ -214,7 +214,7 @@ namespace Octgn.Networking
             counter.SetValue(value, player, false, isScriptChange);
         }
 
-        public void Welcome(ulong id, Guid gameSessionId, bool waitForGameState)
+        public void Welcome(uint id, Guid gameSessionId, bool waitForGameState)
         {
             Program.InPreGame = true;
             Player.LocalPlayer.Id = id;
@@ -224,7 +224,7 @@ namespace Octgn.Networking
             Program.GameEngine.WaitForGameState = waitForGameState;
         }
 
-        public void NewPlayer(ulong id, string nick, ulong pkey, bool invertedTable, bool spectator)
+        public void NewPlayer(uint id, string nick, ulong pkey, bool invertedTable, bool spectator)
         {
             var p = Player.FindIncludingSpectators(id);
             if (p == null)
@@ -263,7 +263,7 @@ namespace Octgn.Networking
         /// <param name="id">An array containing the loaded CardIdentity ids.</param>
         /// <param name="type">An array containing the corresponding CardModel guids (encrypted).</param>
         /// <param name="group">An array indicating the group the cards must be loaded into.</param>
-        public void LoadDeck(Player who, int[] id, Guid[] type, Group[] group, string[] size, string sleeve, bool limited)
+        public void LoadDeck(Player who, ulong[] id, Guid[] type, Group[] group, string[] size, string sleeve, bool limited)
         {
             if (id.Length != type.Length || id.Length != group.Length)
             {
@@ -294,7 +294,7 @@ namespace Octgn.Networking
         /// <param name="type">An array containing the corresponding CardModel guids (encrypted)</param>
         /// <param name="groups">An array indicating the group the cards must be loaded into.</param>
         /// <seealso cref="CreateCard(int[], ulong[], Group)"> for a more efficient way to insert cards inside one group.</seealso>
-        private static void CreateCard(Player who, IList<int> id, IList<Guid> type, IList<Group> groups, IList<string> sizes, string sleeveUrl = "")
+        private static void CreateCard(Player who, IList<ulong> id, IList<Guid> type, IList<Group> groups, IList<string> sizes, string sleeveUrl = "")
         {
             // Ignore cards created by oneself
             if (who == Player.LocalPlayer) return;
@@ -318,7 +318,7 @@ namespace Octgn.Networking
         /// <param name="type">An array containing the corresponding CardModel guids (encrypted)</param>
         /// <param name="group">The group, in which the cards are added.</param>
         /// <seealso cref="CreateCard(int[], ulong[], Group[])"> to add cards to several groups</seealso>
-        public void CreateCard(Player who, int[] id, Guid[] type, string[] size, Group group)
+        public void CreateCard(Player who, ulong[] id, Guid[] type, string[] size, Group group)
         {
             if (who == Player.LocalPlayer) return;
             for (int i = 0; i < id.Length; i++)
@@ -349,7 +349,7 @@ namespace Octgn.Networking
         /// <param name="faceUp">Whether the cards are face up or not.</param>
         /// <param name="key"> </param>
         /// <param name="persist"> </param>
-        public void CreateCardAt(Player player, int[] id, Guid[] modelId, int[] x, int[] y, bool faceUp, bool persist)
+        public void CreateCardAt(Player player, ulong[] id, Guid[] modelId, int[] x, int[] y, bool faceUp, bool persist)
         {
             if (id.Length == 0)
             {
@@ -416,7 +416,7 @@ namespace Octgn.Networking
             }
         }
 
-        public void MoveCard(Player player, int[] card, Group to, int[] idx, bool[] faceUp, bool isScriptMove)
+        public void MoveCard(Player player, ulong[] card, Group to, int[] idx, bool[] faceUp, bool isScriptMove)
         {
             // Ignore cards moved by the local player (already done, for responsiveness)
             var cards = card.Select(Card.Find).Where(x=>x != null).ToArray();
@@ -424,7 +424,7 @@ namespace Octgn.Networking
                 new MoveCards(player, cards, to, idx, faceUp, isScriptMove).Do();
         }
 
-        public void MoveCardAt(Player player, int[] card, int[] x, int[] y, int[] idx, bool[] faceUp, bool isScriptMove)
+        public void MoveCardAt(Player player, ulong[] card, int[] x, int[] y, int[] idx, bool[] faceUp, bool isScriptMove)
         {
             // Get the table control
             Table table = Program.GameEngine.Table;
@@ -673,7 +673,7 @@ namespace Octgn.Networking
             new Rotate(player, card, rot).Do();
         }
 
-        public void Shuffled(Player player, Group group, int[] card, short[] pos)
+        public void Shuffled(Player player, Group group, ulong[] card, short[] pos)
         {
             if (player == Player.LocalPlayer) return;
             ((Pile)group).DoShuffle(card, pos);
@@ -724,7 +724,7 @@ namespace Octgn.Networking
             Program.GameMess.PlayerEvent(player, "hides {0} from {1}.", group, whom);
         }
 
-        public void LookAt(Player player, int uid, Group group, bool look)
+        public void LookAt(Player player, uint uid, Group group, bool look)
         {
             if (look)
             {
@@ -750,7 +750,7 @@ namespace Octgn.Networking
             }
         }
 
-        public void LookAtTop(Player player, int uid, Group group, int count, bool look)
+        public void LookAtTop(Player player, uint uid, Group group, int count, bool look)
         {
             if (look)
             {
@@ -773,7 +773,7 @@ namespace Octgn.Networking
             }
         }
 
-        public void LookAtBottom(Player player, int uid, Group group, int count, bool look)
+        public void LookAtBottom(Player player, uint uid, Group group, int count, bool look)
         {
             if (look)
             {
